# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Garuka API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/cases": {
      "get": {
        "tags": [
          "cases"
        ],
        "summary": "List Cases",
        "operationId": "list_cases",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/CaseStatusEnum"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Level"
            }
          },
          {
            "name": "school_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "School Id"
            }
          },
          {
            "name": "sector_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sector Id"
            }
          },
          {
            "name": "mentor_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mentor Id"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CaseListItem"
                  },
                  "title": "Response List Cases"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cases/{case_id}": {
      "get": {
        "tags": [
          "cases"
        ],
        "summary": "Get Case Detail",
        "operationId": "get_case_detail",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cases/{case_id}/assign": {
      "patch": {
        "tags": [
          "cases"
        ],
        "summary": "Assign Case Mentor",
        "operationId": "assign_case_mentor",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaseAssignRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cases/{case_id}/escalate": {
      "post": {
        "tags": [
          "cases"
        ],
        "summary": "Escalate Case",
        "operationId": "escalate_case",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaseEscalateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cases/{case_id}/notes": {
      "post": {
        "tags": [
          "cases"
        ],
        "summary": "Add Case Note",
        "operationId": "add_case_note",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaseNoteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cases/{case_id}/resolve": {
      "post": {
        "tags": [
          "cases"
        ],
        "summary": "Resolve Case",
        "operationId": "resolve_case",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaseResolveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CaseAssignRequest": {
        "properties": {
          "mentor_id": {
            "type": "string",
            "format": "uuid",
            "title": "Mentor Id"
          }
        },
        "type": "object",
        "required": [
          "mentor_id"
        ],
        "title": "CaseAssignRequest"
      },
      "CaseDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "ref": {
            "type": "string",
            "title": "Ref"
          },
          "student_id": {
            "type": "string",
            "format": "uuid",
            "title": "Student Id"
          },
          "student_name": {
            "type": "string",
            "title": "Student Name"
          },
          "student_gender": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Student Gender"
          },
          "class_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Class Name"
          },
          "school_id": {
            "type": "string",
            "format": "uuid",
            "title": "School Id"
          },
          "school_name": {
            "type": "string",
            "title": "School Name"
          },
          "sector_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sector Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "level": {
            "type": "integer",
            "title": "Level"
          },
          "trigger": {
            "type": "string",
            "title": "Trigger"
          },
          "risk_score": {
            "type": "integer",
            "title": "Risk Score"
          },
          "mentor_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mentor Id"
          },
          "mentor_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mentor Name"
          },
          "opened_at": {
            "type": "string",
            "format": "date-time",
            "title": "Opened At"
          },
          "resolved_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "sla_deadline": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sla Deadline"
          },
          "sla_breached": {
            "type": "boolean",
            "title": "Sla Breached",
            "default": false
          },
          "parent_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Name"
          },
          "parent_phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Phone"
          },
          "metrics": {
            "$ref": "#/components/schemas/CaseMetrics"
          },
          "absence_heatmap": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Absence Heatmap"
          },
          "visits": {
            "items": {
              "$ref": "#/components/schemas/MentorVisitRead"
            },
            "type": "array",
            "title": "Visits"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/CaseEventRead"
            },
            "type": "array",
            "title": "Events"
          }
        },
        "type": "object",
        "required": [
          "id",
          "ref",
          "student_id",
          "student_name",
          "school_id",
          "school_name",
          "status",
          "level",
          "trigger",
          "risk_score",
          "opened_at"
        ],
        "title": "CaseDetail"
      },
      "CaseEscalateRequest": {
        "properties": {
          "to_level": {
            "type": "integer",
            "maximum": 3.0,
            "minimum": 2.0,
            "title": "To Level"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "to_level"
        ],
        "title": "CaseEscalateRequest"
      },
...
      "CaseListItem": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "ref": {
            "type": "string",
            "title": "Ref"
          },
          "student_id": {
            "type": "string",
            "format": "uuid",
            "title": "Student Id"
          },
          "student_name": {
            "type": "string",
            "title": "Student Name"
          },
          "student_gender": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Student Gender"
          },
          "class_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Class Name"
          },
          "school_id": {
            "type": "string",
            "format": "uuid",
            "title": "School Id"
          },
          "school_name": {
            "type": "string",
            "title": "School Name"
          },
          "sector_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sector Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "level": {
            "type": "integer",
            "title": "Level"
          },
          "trigger": {
            "type": "string",
            "title": "Trigger"
          },
          "risk_score": {
            "type": "integer",
            "title": "Risk Score"
          },
          "mentor_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mentor Id"
          },
          "mentor_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mentor Name"
          },
          "opened_at": {
            "type": "string",
            "format": "date-time",
            "title": "Opened At"
          },
          "resolved_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "sla_deadline": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sla Deadline"
          },
          "sla_breached": {
            "type": "boolean",
            "title": "Sla Breached",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "ref",
          "student_id",
          "student_name",
          "school_id",
          "school_name",
          "status",
          "level",
          "trigger",
          "risk_score",
          "opened_at"
        ],
        "title": "CaseListItem"
      },
...
      "CaseNoteRequest": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "CaseNoteRequest"
      },
      "CaseResolveRequest": {
        "properties": {
          "outcome": {
            "type": "string",
            "title": "Outcome"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "outcome"
        ],
        "title": "CaseResolveRequest"
      },
      "CaseStatusEnum": {
        "type": "string",
        "enum": [
          "open",
          "mentor_assigned",
          "visited",
          "escalated_sector",
          "escalated_district",
          "resolved_returned",
          "closed_moved",
          "closed_other"
        ],
        "title": "CaseStatusEnum"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}